nvvar: Fix creation without value
When a new nvvar is added without a value we want to get the value
from the corresponding globalvar. We do this by using nv_set which
ends up freeing the exact string that we passed into nv_set as
value. This results in a corrupt value. Fix this by assigning a
value manually without calling nv_set.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>