param: make parameter functions more consistent
This patch creates a consitent set of device parameter functions.

With this we have: dev_add_param_<type><access>

"type" is one of: int32, uint32, int64, uint64, string, mac, ipv4, enum, bitmask
The improvement here is that we now can exactly specify the width of the
int type parameters and also correctly distinguish between signed and
unsigned variables which means that a variable no longer ends up with
INT_MAX when it's assigned -1.

"access" can be empty for regular read/write parameter, "_ro" for readonly
parameters which get their value from a variable pointer in the
background or "_fixed" for parameters which are set to a fixed value
(without a pointer in the background).

Some more exotic types are not (yet) implemented, like
dev_add_param_ip_ro.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>