libgen: implement posix_basename
There are two different versions of basename(): The GNU version and the
POSIX version. The GNU version never modifies its argument and returns
the empty string when path has a trailing slash, and in particular also
when it is "/". The POSIX version modifies its argument and thus works
properly with strings which have a trailing "/".

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>