ARM: i.MX50: Add imx50.dtsi as workaround
The upstream imx50.dtsi is missing the usbphy and usbmisc nodes.
Furthermore, the clock property values for the two existing usb
instances (usbotg and usbh1) is set to a clock which is not present
on the imx50.
Note that usbh2 and usbh3 do not exist on the imx50.

Fix imx50.dtsi to get the usb peripherial mode via DT running,
until imx50.dtsi is fixed upstream.

Signed-off-by: Alexander Kurz <akurz@blala.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>