docs: fix hiding of chapter from imx.rst
The files 'mxs.rst' and 'imx.rst' used the same chapter headline:

    Freescale i.MXs

This confuses sphinx. It does not support having the same headline at
the same level twice and hides the other chapter. Just rename the
chapter heading to avoid that.

And rename the subchapter headline to 'MXS boards' for clarity.

Signed-off-by: Stefan Lengfeld <contact@stefanchrist.eu>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>