usb: imx: Only use clock when it exists
Fixes: bbe0df90b8e2463dd8db651e9f1dd57702985d2d

I assumed all chipidea devices actually have a clock. This is true
when looking at the device tree, but in barebox not all i.MX SoCs
use the clocks provided by the device tree. Most instead use
clkdev_add_physbase() to attach a clocks to devices. For these the
chipidea device does not have a clock.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>