cdev: make file operations const
scripts/checkpatch.pl warns that struct file_operations should be const,
but cdev->ops is not const, so without this patch we can choose between
a warning from checkpatch and a warning from the compiler about
discarding the const attribute when assigning the struct file_operations
cdev->ops.

Since there is no reason to modify the contents of cdev->ops after
probing, make it const.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>