ratp: fix single byte sending flagged with SO
When a single data byte is sent, it is flagged as SO and the actual
data goes in the length byte within the header.

Worth noting that this issue wasn't found in any barebox<->bbremote
interaction because all data packets sent between them always have
more than one byte (due to the barebox command specific header).

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>