state: backend: always pass backend type to Linux dtb
When the "backend-storage-type" property is unset in the barebox dtb
then barebox will use the default. Instead of leaving the property
unset in the Linux dtb, set it to the value we used as default. By
making the default explicit to Linux we force Linux to the same backend
type, even when the defaults may differ.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>