readlink: Improve -f handling
For `readlink -f FILE VARIABLE`, set VARIABLE to the absolute, canonical
file name of FILE by following symbolic links.  All but the final
component of FILE must exist.

Prior to commit a602bebcf7e4 ("fs: Implement links to directories") the
above worked in the limited case where FILE was an absolute path, the
final component was a symbolic link, and all preceding components were
directories.  If FILE was a relative path with the final component being
a symbolic link, and all preceding paths being directories, the command
sort of worked, but produced a relative path as the result.  Both of
these cases were broken by the above commit.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Lucas Stach <l.stach@pengutronix.de>