scripts: imx imx-usb-loader: Fix last transfer error message
When doing memory read transfers there will always be 64 bytes
transferred, even when less bytes are requested. This is expected
and there is a test skipping the error message in this case. The
test is wrong though since cnt is not decremented and will never
be equal to rem. Fix the test so that verifying memory does
not give a bogus error message.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>