usbgadget: fastboot: Allow to automatically export the bbu handlers
We have a list of registered handlers which take a barebox update. Do
the next step and allow to automatically export them via fastboot so
that barebox can be updated via fastboot without manually exporting
the partition. Since this may not be desirable in all cases this
behaviour is configurable.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>