e1000: implement support for smaller flash chips
Even though the register description of the FLA register suggests (on
page 386 of the i210 datasheet (rev. 3.1)) that 2 MB, 4 MB and 8 MB are
the only supported flash sizes, the list of supported flashes (on page
794 of the same document) lists for example a Micron M25PE80 which only
has a size of 1 MB.

Also in general it seems sensible to stick to the formula given for the
meaning of the values that are not explicitly listed.

Acked-by: Andrey Smirnov <andrew.smirnov@gmail.com>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>