arm: bootm: don't fall over if image is padded with less than 40 bytes
If the zImage has a padding, which is less than 40 bytes
(sizeof struct fdt_header) the amount of read bytes would be propagated
as an error code. Fix this by only propagating real errors and treating
failure to read less than the expected amount as no concatenated DT
being present.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>