of: base: use root_node compatible as suggestion for a hostname
on some SoCs we can use generic PLL and RAM initialization. In this
cases we create board file only to provide a host name.
With this patch host name will be created from device tree compatible.
For example:
 compatible = "board_vendor,board", "chip_vendor,soc"
the host name will be:
 "board"

This function will not overwrite a host name which is already set by
board or machine code.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>