Documentation: change default highlight language to shell
There are many places where sphinx complains that blocks cannot
be hilghlighted in 'c' and in fact I am not aware of any C code
snippets in the documentation, so change the default highlighting
style to shell which fits our needs much better. With this we
get rid of many warnings while building the documentation.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>