ARM: dts: bcm2835-rpi: re-enable booting from SD card
Kernel commit a19adf8d86e822eba502486524127595372c85f6 [0] changed the
RPi1 device tree from &sdhci to &sdhost, which was imported in barebox
commit d14b844b08635c717fb52a294ed8d6872e260315. It seems that barebox
does not have a driver for brcm,bcm2835-sdhost (yet), which resulted in
barebox unable to boot from SD card. However, the driver for
brcm,bcm2835-sdhci worked fine in the past.

Although the upstream change was made for good reasons, the simplest
change for barebox for now is to revert to using &sdhci and disable
&sdhost.

Thanks to Alexander Dahl for raising this issue on the DistroKit
mailing list!

[0]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=a19adf8d86e822eba502486524127595372c85f6

Cc: Alexander Dahl <post@lespocky.de>
Signed-off-by: Roland Hieber <rohieb@rohieb.name>
Reviewed-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>