file_list: Add ubi flag
This flag is added to better support flashing UBI image with Android
sparse images. Android fastboot splits images which are bigger than the
free memory into multiple images which are then transferred in multiple
fastboot sessions. In a session which is not the first one we can no
longer detect if an image is a UBI image or not, so we need a flag to
make this explicit.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>