ARM: socfpga: move environment from dtsi to board dts
Currently all upstream boards use the same barebox-environment setup.
It is however problematic to add new boards, that have a different setup.
Instead of having to /delete-node/ the file-path property in those boards,
move the environment to the specific board files.
This also has the charme of - maybe one day - getting rid of a barebox-version
of socfpga.dtsi :-)

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>